# ⚡️ Speed up function `default_pt` by 54%