--- a/litellm/litellm_core_utils/prompt_templates/factory.py
+++ b/litellm/litellm_core_utils/prompt_templates/factory.py
@@ -40,7 +40,8 @@ from .image_handling import convert_url_to_base64
 
 
 def default_pt(messages):
-    return " ".join(message["content"] for message in messages)
+    # Use list comprehension for better performance over generator in join
+    return " ".join([msg["content"] for msg in messages])
 
 
 def prompt_injection_detection_default_pt():